// src/app/models/credit.ts
export interface Credit {
  idPerson: number;
  namePerson: string;
  lastNamePerson: string;
  numberIdentificationPerson: string;
  idTypeIdentification: string;
  nameIdTypeIdentification: string;
  idCredit: number;
  totalCost: number;
  totalFee: number;
  idCity: number;
  nameCity: string;
  idSuburb: number;
  nameSuburb: string;
}